Update the included Gruntfile to use Webpack to	perform	SASS compilation.
This CR	adds to the existing static Gruntfile. Now, using Webpack, we can
perform	SASS compilation of whatever is	in the static/scss directory. For
now, there is just a placeholder SCSS file that is empty.

TEST=manual

Added some actual SCSS files to	the static/scss	directory and saw that they
were correctly compiled	to a single CSS file using Grunt/Webpack.