Provide some placeholder JS in the static directory.
This PR provides an example of how one would include some additional JS in the
theme's static directory. All JS and SCSS is imported/used through a single
entry point, called entry.js. Right now, the JS directory contains a single
file that exports a function to print 'Hello World'. I added the Closure
compiler plugin to Webpack. That way, we use Closure to minimize any customer
supplied JS. I also added the CleanWebpackPlugin. This plugin cleans out the
dist directory as the first step of a Webpack bundle. This is the best
practice.

TEST=manual

Made sure that the JS was bundled properly and the CSS was too. Created a
simple HTML page to verify that 'Hello' was printed to the console when using
bundle.js and that the bundle.css could also be correctly imported.