Add additional functionality to Gruntfile and the Webpack config.
This PR adds new functionality to the Gruntfile and Webpack config:
     - There is now a 'watch' Grunt task that listens for changes to various
       directories in the repo. If a change is detected, Jambo 'build' is run.
     - The 'webpack' Grunt task can now load/bundle static assets such as fonts
       and images.
These changes were initially required as part of our work on the First
Financial test site.

TEST=manual

Tested the updated Gruntfile and Webpack config with the First Financial test
site. Made sure that the 'watch' task worked as expected and that fonts and
images were bundled with the CSS.