Add a sample HTML layout to the theme.
This PR creates a simple layout to be included with the theme. This layout
suggests how HTML pages generated from the theme's templates might be
structured. Ultimately, the contents of the layouts directory will be copied
into the top-level repository's partials directory on import. This will allow
a user to easily change the included layouts.

TEST=none