Incorporate DOM and SASS from AEB and the template-answers repo.
This PR updates the page.html.hbs of the existing templates to have the DOM
structure expected by the AEB styling. The styling from template-answers and
the AEB is now included as part of the HH Theme's static assets. I also updated
the script partial for UniversalResults to account for recent changes made to
v0.14.

I looked at the AEB source to determine how to structure the DOM in the
page.html.hbs files. All class new class names were taken from AEB. The added
SCSS is essentially just a copy of what is in template-answers. I did have to
make some tweaks.

TEST=manual

I made a vertical and universal site and ensured that the AEB styling was
getting applied. Also made sure the DOM structure matched existing sites
built with AEB.