Update Partial naming scheme in accordance with Jambo changes.
The Partial naming conventions in Jambo have recently been changed. Now, Jambo
names a Partial according to its relative path, with the file extension
removed. The Theme must be updated to reflect this change.

J=SPR-1988
TEST=manual

Made a simple test site with the updated Theme. Made sure it could Jambo build
without errors and that it looked as expected when served.