Include all card attributes in dataForRender for Standard and Accordion.
This PR ensures that both Standard and Accordion cards list all available
attributes in their respective dataForRender method. Those attributes that do
not have a set value are left commented out to demonstrate how they could
be added.

Also included in the PR is a fix for the HH Theme. Jambo recently updated the
naming scheme for Partials. The HH Theme needed to be updated to accomodate for
the differently named partials. There was a PR with the HH Theme updates, but
I missed a couple of partial references.

J=SPR-1985
TEST=none