Move from SDK v0.14.0 to v0.13.3 (#36)
v0.14.0 will not be released for a while; v0.13.3 is going to
be v1.0.0 when it is released. The HH theme should
be tested with v0.13.3 rather than v0.14.0.

TEST=manual

jambo build and inspect page in the browser to verify that the script/link tags
were for v0.13.3. Briefly test Universal and Vertical page by clicking around and seeing
no console errors.