Support 'Show more'/'Show less' toggling for Hitchhiker cards. (#53)
This PR adds support for the showMoreLimit to the Theme's base card_component class. By adding a few lines to its HBS template and exposing a couple of configuration options, any Card can now support this toggling behavior. As part of this PR, I made it so the StandardCard supported it. There is also common, shareable SCSS to style the toggle button.

J=SPR-2057
TEST=manual

Verified the following for the StandardCard:

- Toggling worked as expected when showMoreLimit was specified and the card's
  details were longer than the limit.
- If a showMoreLimit was added, but the details were shorter than that, no
  toggling was added.
- Things worked as expected when no showMoreLimit was specified.
- showMoreText and showLessText worked as expected.
- Styling of the toggle looked good.