Various updates to the StandardCard component.
This PR makes various updates to the HH Theme's StandardCard:

- If there is no url for the title, style it as normal text.
- Default title url's target to _top.
- Remove callsToAction array and replace it with two, hard-coded attributes:
  CTA1 and CTA2.
- Default title's url to landingPageUrl if websites is not present.

J=SPR-2057
TEST=manual

Manually verified all four updates above. For the CTA updates, I made sure that
CTA1 and CTA2 rendered as expected when both were specified and had urls. I
also made sure that things rendered as expected if either was missing or if
one did not have a url.