Add 'any' and 'all' handlebars helpers (#54)
Add 'all' and 'any' handlebars helpers that take in a variable amount of arguments. Adding these handlebars helpers allows multiple conditions in 'if' statements. Usage for that would look like this:

{{#if (any A B) }}

TEST=manual

On local site, test compilation of handlebars templates using both 'all' and 'any'. Set A to true, B to false; A false, B false; A true, B true. Verify that for 'all' and 'any' that the statements evaluate properly.