Config: Set optional layout config to empty string (#46)
Updating optional layout config to empty string. If HH's do not want to enter anything for this config, now no action is required. Previously, a HH would have had to delete "" or the entire config option.

TEST=manual

Generate page of each type, note that layout elements corresponding to this config are not present in the DOM.