Add global 'Formatter' alias for HitchhikerJS.Formatters class.
This PR adds a 'Formatter' alias to the window that points to the Formatters
class exported by HitchhikerJS. This was requested by Liz to more closely
match the behavior from AEB. I updated the LocationCard to use the new alias.

J=SPR-2127
TEST=manual

Made sure the urls for both LocationCard CTAs were formatted properly.