Add Event Card (#59)
Add a new card type to the HH theme. The Event Card has a big date on the left side, title, subtitle, description, and one CTA by default. 

TEST=manual

Use the Event Card on a Vertical and Universal page. Test on mobile, tablet, and desktop. Hardcode data in dataForRender to test no data and long data cases.