Map Component to default to config from verticalsToConfig if present (#79)
If there is a map config in the page's verticalsToConfig (for the verticalKey of the page), then the Map Component will default to using that configuration. If there is Map configuration provided in component settings, that will still take priority.

TEST=manual

Use locator template with map config in componentSettings, verticalsToConfig, both, and neither.