Accordion Card optionally expanded or collapsed on load (#84)
Move the height: auto below the height: 0 for CSS specificity so content shows on page load if option is true. Move isExpanded variable into dataForRender instead of a separate setState.

TEST=manual

Use AccordionCard on a test site and test with isExpanded true and false. Test with missing isExpanded to ensure default is correct.