Universal Template: Accept vertical config from universal page's config.json (#76)
This allows vertical config for the Universal page to live in the Universal page's config. This may be used for verticals that don't have a page or in cases where configuration is different on the Universal page than on the Vertical page. If the Universal page's config has information for a vertical, it will not attempt to pull any config from the Vertical's page (if it exists).

TEST=manual

Generate a Universal page. Add configuration for a vertical without a page. Add (different) configuration for a vertical with a page.