Fix extra parenthesis in getDefaultMapApiKey and remove from vertical results template I was almost very upset at myself when I noticed this but apparently you're allowed to have extra parenthesis in javascript (???) so I guess I can test code fine just not write it (which is better than neither being true)
defaulting the map api key doesn't work on the vertical results template anymore because we
don't do a deep merge of the config in the hbs and the config from the
json, and mapConfig is an object. It still works in universal and locator (tested
by deleting apiKey everywhere and still seeing maps after jambo build)
This should be fine, though, since we discussed that people should just be using
the locator template if they want a map anyways.