Add Sort Options to Vertical-Standard and Vertical-Map templates (#88)
Add partials for the SortOptions component to the Vertical-Standard and Vertical-Map page templates.

TEST=manual

Add sort options config to a page built with the Vertical-Standard page template. Generate page with Jambo and see Sort Options component. Test sorting and ensure that behavior is the same on a test QA site built without Jambo. Do the same for a page built with the Vertical-Map page template.