Add support for a conversionTrackingEnabled global_config option. (#103)
This PR adds support for a conversionTrackingEnabled option in the
global_config. When true, this option will ensure that conversionTracking
is enabled in the SDK and the needed ytag script is included. Note that the
option defaults to false, as discussed with Liz.

J=SPR-2139
TEST=manaul

Set the new config option to true and obeserved that analytics events were
reported to the correct endpoint and that they also included a first party
cookie. Removed the config option and observed analytics events did not
include such a cookie and were reported to the correct endpoint.