Add runnable iFrame script, with resizing
This PR is to match AEBs current situation with iframes.

Create a script that will initialize an iframe pointing to the answers
experience of the repository (iframe.js). Dynamically resize the iframe
height based on the content in the answers experience. Make the iframe
responsive to query parameters, to link to specific pages within an
iframe.  IFrames are created by default. The feature is opt-out through
an environment variable.

NOTE: This is only for en locales, we do not have the localization
strategy in place yet.

J=SPR-2033
TEST=manual

Create a test site. Try the following build commands:
        STAGING_DOMAIN="http://answers.yext.com/" PROD_DOMAIN="http://answers.yext.com/" grunt webpack
        JAMBO_INJECTED_DATA='"{HIDE_IFRAME: false}"' jambo build

This should create a file iframe.js in your /public folder. If you
create a static page referencing this script and add an
"#answers-container" div to the page, you should see the answers
experience  as an iframe.

If you do a normal (without environment variables)
        grunt webpack
        jambo build

You will still get an iframe.js file, it will not function correctly.