Use transform data to enable universal result limits (#106)
The AEB supported this, so for parity the HH theme needs to support it too. The back end of Answers does not currently support result limits on universal results, when that work is complete we should update this.

TEST=manual

On universal page, specify a limit through vTC. See results limited to the number I specified.