Add showOrdinal data mapping to the location-standard card. (#113)
Add showOrdinal data mapping to the location-standard card.

This new data mapping will default to true. When enabled, it displays the ordinal on a location-standard card. I noted an issue with the SDK where a card's ordinal does not take the search offset into account. This will be fixed separately.

J=SPR-2204
TEST=manual

Set showOrdinal to true for location-standard cards with and without images. Compared these two the images provided by Rose for guidance.