Fixes for ifeq 'true', add quotes, add missing param to universal results (#123)
Previously the theme was using {{#ifeq isFirst 'true'}} checks to test if the isFirst variable was true. That check only worked if the JSON's isFirst value was the string "true", but did not work with the boolean true. Updated to check if the isFirst config option is present, if so, use it. Since there are no quotes in the hbs, "true" as a string or as a boolean would be printed without the quotes.

Add quotes to the viewAllText in UniversalResults, they were previously absent which caused parse errors.

Add includeMap param to UniversalResults if the user passes it through.

TEST=manual

Use local test site, specify isFirst using boolean then string for true and false values. Add viewAllText to a vertical in verticalsToConfig. Add a map on UniversalResults using the includeMap config option.