Assorted map configuration updates requested by the HHs. (#136)
This PR includes a couple updates around the map configuration that were
requested by the HHs:

- For the universal template, if a VTC object lists a mapConfig for a vertical,
  we should interpret that vertical as having includeMap: true.
- The vertical-map template's JSON file should add mapConfig under VTC instead
  of having it under componentSettings.Map. This will encourage HHs to always
  use VTC for map configuration.

TEST=manual

Created a page with the updated vertical-map template. Ensured it looked as
expected. Created a universal page. Saw a map for a vertical that had mapConfig
set in it's VTC. In the universal config's VTC, added a new vertical with a
mapConfig. Saw a map for that vertical on the universal page too.