Add vertical-grid template
This template should be exactly the same as VerticalStandard except we
use the Product Prominent Image by default and have the vertical results
maxNumberOfColumns set to 3 by default

TEST=manual

Tested on a fresh jambo site with universal (should show no grid cards)
and a vertical page (based off of the vertical-grid template). Checked
to make sure multiple grid columns could be added. Checked to make sure
the grid worked well with facets (currently the width of the page is not
wide enough to allow for gridding when there are facets). Check to see
the card borders on only the content not the whitespace.