Add results title bar variables by default, fix --yxt background var (#161)
From the HH: "These are helpful when we want to use the brand's primary
color as the background for the title bar on Universal Results."

Fix --hh -> --yxt variable

J=SPR-2335
TEST=manual

Tested on a jambo site with FAQs, re-imported files from the theme
Changed the new variables to make sure we saw changes in the results title
bar. Checked to see FAQ background color on hover matched the
background-hover variable.