Default title link for job standard card to landingPageUrl
It seems the Job entity does not have the website field. Change the
default url to the landing page url so the HHs do not have to fork the
card for every job standard.

J=None
TEST=manual

Created Job entity with two urls:
        Application URL: http://www.yext.com/apply
        Landing Page URL: http://www.yext.com/landing

Added to rosetest answers experience

Tested on a new site to see that the url for a job entity
is using the landingPageUrl instead of the (non-existant) website field