Changed faq-accordion card to use rtf answer field (#169)
Updated the faq-accordion card to call the RTF formatting function for the answer field.

J=SPR-2348

TEST=manual

Test on desktop by testing on a local site with faq vertical. Inserted a custom RTF for an faq entity and used dataForRender method to test if the styling of the RTF is correct.