Direct Answer card scaffolding, add base card component (#176)
Scaffolding for Direct Answer Card base component and the default card. The dataForRender function will change in a later PR to support answer.value's that are lists. This PR is to define the general structure of the Direct Answer cards.

TEST=manual

Test with the dev/direct-answers-card branch of the SDK and log data to see that the expected data comes through. Add "{{> directanswercards/all }}" to the page.html.hbs, and add the "defaultCard" config option to the page config.json. Run npx jambo build && grunt webpack and view universal page in browser. Search for a term that returns a direct answer, see (local, removed for the PR) console.logs printing expected data, and the "Placeholder" HTML display.