Use TerserPlugin for minimization in the Webpack task. (#178)
* Use TerserPlugin for minimization in the Webpack task.

This PR updates the Webpack task to use the TerserPlugin when minimizing the
JS bundle. The TerserPlugin is used by default in Webpack. Previously, we used
the Closure compiler to optimize our JS. However, the compiler's plugin was
slow to run. On average, the Webpack task took about 10 seconds when Closure
was used. The task takes, on average, 6 seconds with the TerserPlugin. What's
more, the TerserPlugin seems to provide a smaller sized bundle than Closure.

J=SPR-2407
TEST=compile

Made sure the Webpack task produced the JS bundle successfully when using the
TerserPlugin for minimization.