add default files for answers-variables and fonts in theme (#173)
These files are meant to be defaults that are subject to change in the
theme. Thus if we add more variables to the theme, we will not touch the
repo-level variable files. These new variables will be behind the scenes
and repo-level work on styling will continue to work.

J=SPR-2372
TEST=manual

Tested on a new Jambo site. Tested a variable change in both the
theme-level "-default" files and the repo-level HH-editable files. Made
sure the repo-level variables took precedence but when not present, the
defaults showed up.