Update iframe.js for referrerPageUrl defaults
When creating an iFrame, we should pass the referrerUrl to the iframe
source URL. If it does not exist in the parent URL, then we should use
document.referrer.

J=SPR-2452
TEST=manual

Tested on a QA site to test the iFrame navigation.

Test referrerPageUrl from parent is in iframe src URL.
Test referrerPageUrl is document.referrer if parent URL does not have a
referrerPageUrl query param