Keep static assets in the output dir folder for HH custom images
We have no way of loading file paths if they come from the HH config
files. In order to allow HHs to reference static assets with

"static/assets/images/file.svg

in config files, we should keep the containing folder in the desktop
folder.

Note: This is a short-term solution. In the future, HHs will keep this
interface of referencing "static/assets/images/file.svg" except we will
move the desired files to the top-level desktop as we do with other
assets (e.g. footer-logo.svg) and replace their references.

J=SPR-2456
TEST=manual

In a config/index.json add to the componentSettings.SearchBar
      "customIconUrl": "static/assets/images/test2.svg"

Check to see the asset is in the output dir and the search icon is
updated.
Check to see the other assets are working as expected (fonts, scss,
footer-logo.svg)