Add image loading and hashing for json config in html
Previously, if an admin added a path to a local image asset (e.g.
static/assets/image/test.svg), to a *.json file, then we would need to
keep that exact path in the output folder. This change the images in two
ways:
* Images are replaced with paths referencing the top-level out (e.g.
  instead of desktop/static/assets/image/test.svg, we get
  desktop/test.svg)
* Images are hashed (e.g. instead of desktop/test.svg, we get
  desktop/test.<contenthash>.svg)

HHs will still input the path has `static/assets/image/test.svg`,
webpack handles the path resolution using a new custom html loader.

Note: this must be run after the HTML loader, as it relies on the
following (1) the html has been converted to javascript (2) the double
quotes are escaped (for the regex step) (3) an html-loader helper
function is available

The implementation is heavily influenced by the html-loader
implementation. For example, the import syntax is copied nearly directly
(simplified for our use case):

https://github.com/webpack-contrib/html-loader/blob/master/src/utils.js#L60

Note: image paths that are in img src tags are handled by the original
html-loader. This is for img srcs that werereferenced in the JSON.

J=SPR-2456
TEST=manual

Test site with an updated local Jambo command to import the
webpack/html-image-loader.js

On a local Jambo site in config/index.json componentSettings:
Put test2.svg in static/assets/images
```
    "SearchBar": {
      "placeholderText": "Search", // The placeholder text in the answers search bar
      "customIconUrl": "static/assets/images/test2.svg"
    }
```
Verify the searchbar icon is the test2.svg
Verify the image src in the DOM has a content hash

Put test2.svg in multiple places in the code (e.g. sectionTitleIconUrl).
Verify all imports work (you visually see the different svg).

Verify the yext-logo.svg still works (used in HTML, not in JSON
configuration).