Add answers.js as an alias for the iframe.js (#233)
Legacy repositories under AEB would use answers.js as the script to send
to clients. As a result of changing to iframe.js, we should make a
duplicate answers.js that is a copy of iframe.js so clients can refer to
the same file.

J=SPR-2598
TEST=manual

Check an iframe shows with the new answers.js that is the exact same as
the iframe.js experience.