Bundle iFrameResizer library with the various iframe assets.
The iFrameResizer library is now included with all generated iframe assets.
This was done to support a customer who was unable to fetch the library from
Cloudflare because of CSP issues. Note that the html.hbs partial still fetches
the iFrameResizer.contentWindow library from Cloudflare. This should be fine,
however, since the iframe'd pages will not be subject to the same CSP issues.

J=SPR-2617
TEST=manual

Applied these Theme changes to a local Jambo repo. Verified that the iframe'd
experience was getting resized when queries were made. Also verified that the
URL params from the iframe were communicated to the parent.