Fix bug that ignored componentSettings.Navigation configuration. (#246)
The Theme expected the NavigationComponent's settings to be located in
componentSettings.navigation, as opposed to componentSettings.Navigation. This
was confusing to the HHs since all other entries in componentSettings are
capitalized. This issue was present in the navigation script partial for all
page templates.

TEST=manual

Made a page from each type of template. Added componentSettings.Navigation
configuration. Saw that configuration applied correctly on each page.