Fix resolve-url-loader plugin for SASS relative paths (#236)
The webpack resolve-url-loader plugin was not previously working. This is the plugin that allows paths to be relative from the file they are in (rather than the SCSS entry point file). Per the resolve-url-loader documentation: "source-maps required for loaders preceding resolve-url-loader". Added the source map option to the sass loader, and the resolve-url-loader plugin is working.

TEST=compilation