Add upgrade theme script
This commit adds the upgrade theme script, which is
meant to be run after a jambo upgrade.
jambo upgrade updates the themes/answers-hitchhiker-theme
folder, or whatever the default theme happens to be, but
does not perform any further updates to the repo. What
needs to happen after that point is completely theme specific.

This script can take a --isLegacy command line argument, to specify
whether to perform a legacy or standard theme upgrade (mirroring
current fambo behavior). It was necessary to make this a node
script instead of a bash script for access to tools like CommentJson.
(as well as just being easier to write and read this way).

This file was changed from a normal to executable file so that it
can be invoked with the ./ syntax

J=SPR-2631
TEST=manual

tested that root level files are copied correctly (e.g. package.json things)
tested that global_configs are merged the same way fambo would merge it
(ran a few side by side tests)
test that fonts.scss file is correctly copied and moved from anywhere
in the static/ folder to static/scss
test that static/assets/images are preserved except for yext-logo.svg
test that static/assets/fonts are preserved except for the 3 default fonts.