Replace the CleanWebpackPlugin with the RemoveFilesPlugin. (#268)
The CleanWebpackPlugin was deleting anything in the output directory that was
not altered by Webpack in some way. This included things listed in Jambo's
preservedFiles. I was able to configure the plugin in such a way that this
behavior was prevented. But, it felt brittle to me. Additionally, this plugin
is marked as dperecated and no longer maintained. So, I replaced it with the
RemoveFilesPlugin. The plugin is configured to delete the 'static' directory
from the output after every Webpack compilation. It deletes nothing else.

J=SLAP-11
TEST=manual

Performed a manual Webpack compilation and verified that the output directory
looked as expected (with preservedFiles intact). Also performed automated
compilations via the 'grunt watch' task. Verified that the output directory
looked as expected after each of these.