Add universal grid options three-columns and two-columns (#275)
Add grid options for the universal standard template. This is to support a grid
of result cards not only in the vertical results but also the universal
results. We start with two and three columns set at a specific width from the
spec. On mobile there is no grid and we use the standard one column with no
padding.

Implementation detail: we move out universal standard into a mixin to DRY up
the sass files.

Example usage

```
  "verticalsToConfig": {
    "products": { // The vertical key from your search configuration
      "cardType": "product", // The name of the card to use - e.g. accordion, location, customcard 
      "icon": "tag", // The icon to use on the card for this vertical
      "verticalLimit": 12,
      "universalSectionTemplate": "three-columns"
    }
  }
```

J=SLAP-505
TEST=manual

Test on a universal page using the above example as inspiration. Test
only one vertical with two-results. One one vertical with three-results.
Two verticals with each. Using both two-results and three-results on
the page. No changes should be done to the standard universal results
section template. Make sure all cards have the set width. Sanity test
on IE11/firefox. Test verticals with maps on universal and vertical. Make
sure cards on vertical are still working with the new height/width specification.