Use quotes instead of apostrophes to allow apostrophes in JSON config
Since we're putting raw configuration directly into a string bounded by
single quotes, "'", single quotes in the configuration (e.g. apostrophes)
cause JS syntax errors. To fix this, we've opted to switch to quotation
marks. Some of the changes in this PR are specifically to resolve the
configuration issue, and others are to ensure that the code is uniform to
make it more intuitive to continue to use quotation marks as we maintain
this code.

TEST=manual

Generate a page with each of the four templates: vertical-standard,
vertical-map, vertical-grid, universal-standard. Add every component
and add config for each config option mapped by the theme, using
apostrophes. Run npx jambo build && grunt webpack && serve, and
view page in browser without JS errors.