iframe: Use replaceHistory in onStageChange to fix facet loop bug (#290)
* iframe: Use replaceHistory in onStageChange in v1.5.3 to fix facet loop bug

The iframe script does not know when to replaceHistory when change
browser history to mock the SDK changes. Previously, we would guess with
replaceHistory using checks for context/referrerPageUrl. However, this
has become a problem with facets. With SDK v1.5.3, we are exposed to
replaceHistory in onStateChange. This allows us to remove our previous
checks and reliably change state whenever the SDK changes state.

J=SLAP-577
TEST=manual

Test a local iframe script on a new Jambo HH Theme site using SDK
v1.5.3. Create an iframe_test.html that references iframe.js, remember
to build the site with the correct injected data for the iframe URLs.

* Remove default messageData.message