Properly format RTF direct answers in the allfields-standard card. (#294)
RTF direct answers were not properly formatted. There was a missing 'break'
in the 'rich_text' case which led the code to fall through to the 'default'
block. The default block overwrote the correctly formatted value. Lastly,
we were not including the needed CSS styling for when there was an RTF direct
answer.

To properly apply the styling, I had to add a 'HitchhikerAllFieldStandard-field
ValueRtf' class. I couldn't just add the styling under the 'HitchhikerAllFields
Standard-fieldValue' selector. That would have caused the 'fieldValueLink'
styling to be break as it styles links a bit differently than the RTF mixin.

T=345108
TEST=manual

Made sure the formatting/display was correct for an RTF and a phone # direct
answer.