formatters: _formatHoursForAnswers does not have side effects (#300)
This is to fix a bug where if you pressed the feedback button on a
direct answer card, there would be a console error saying replace is not
a function. This was because we were parsing the time intervals into
numbers and changing the input parameter (the answer). When a user tried
to click on the feedback button, the parameter was already modified into
a number so when it tries to re-render the data, it will try to do a
.replace on a number (not a valid function for the number type). We
change things to not have side effects and instead return a new /deep/
clone of the input answer.

J=SLAP-608
TEST=manual

Test on an experience with a Direct Answer w/ Hours. (Used the Yext
                Answers experience with the query "ny hours").

Test that the direct answer can be seen correctly and when clicking on a
feedback thumbs up input on the direct answer card, the console does not
error and you see the "Thank you for your feedback!" copy.

Test that other direct answers are fine as well (address), even though
this should be isolated to hours direct answers.