formatters: Add isTwentyFourHourClock option to openStatus (#301)
We add the option for isTwentyFourHourClock to the formatter for
internationalization. This will allow a user to call openStatus and
receive a formatted string back with the time string in 24 hour time.
For example "Open Now - Closes at 18:00". Previously we were hardcoding
this to false.

J=SLAP-607
TEST=manual

Test that this openStatus(profile, 'en-US', true) would give you
international 24-hour time and that this works for both true and false
on a local HH theme site.