bundle: Add polyfills and plugins from SDK (#278)
We removed global polyfills in the SDK. The theme was relying on these global
polyfills for browser support. We add the global polyfills from the SDK to the
theme for now. This is with the context that the theme is normally controlled
by Yext. This is also with the context that when clients use the theme on their
sites, it is normally within the sandbox of an iFrame.

Note: I tested that iFrames do not affect the implementations of the parent
with the following test. In entry.js:
```
Object.assign = () => {
  console.log('hi');
}
```
In console of the experience, `Object.assign({}, ['a','b']);` gives a hi.
In console of an `iframe_test.html` that opens the experience in an iframe,
`Object.assign({}, ['a','b']);` does not give a hi.

J=SLAP-555
TEST=manual

Test basic functionality for a Jambo site referencing a local SDK dist (without
global polyfills) in Edge, IE11 on Windows. In Firefox, Chrome on MacOS. 
On Safari on iOS.