Formatters: fix typo cloneDeep to clonedeep
Fix small typo in lodash.clonedeep import. This typo
caused the webpack build to fail on the live preview images,
but not locally on a mac. This is probably because mac osx
is case insensitive while unix-style systems are case sensitive.

TEST=manual

run "npx jambo build && grunt webpack" locally
run "grunt watch" locally and edit a config file
add this change to a site through sgs and see live preview
  + the custom development build both work