Add vertical-map-filters page template
This commit adds the code for the vertical-map-filters page
aka collapsible filters. It creates 2 new theme components,
which are in the top level collapsiblefilters folder, and
also adds special config to a number of components. It also
performs some additional setup in the answers onReady(), which
is detailed in the page-setup.hbs file, and registers some
functions within the onReady(). In the future we may need
some way to not "dirty" the scope in the on ready, on the
off chance that we happen to use the same variable names as some
other arbitrary code in the onReady(), but for now this
shouldn't be an issue.

TEST=manual

create a new jambo repo, import the theme, change the theme
to this branch, create a new collapsible filters page and see
that it works both on mobile/desktop, and also on ie11/chrome/safari/firefox.

I also have some testcafe acceptance tests, but we don't have good infrastructure
in the theme for these yet, so these have not been added yet.